Implement cfpq using pygraphblas
