+++ project/with_scipy/bool_decomposed_nfa.py
@@ -8,7 +8,7 @@ from typing import Union
 import pyformlang.finite_automaton
 import scipy
 
-from project.scipy_sparse_bool_decomposition import BoolDecomposition
+from project.with_scipy.bool_decomposition import BoolDecomposition
 
 __all__ = ["BoolDecomposedNFA"]
