+++ project/with_scipy/bool_decomposition.py
